+++ serde-query/src/lib.rs
@@ -45,6 +45,9 @@
 //! `serde-query` currently supports the following syntax for accessing a part of the input.
 //!
 //! * `.field` for accessing a field with a name `field` of an object.
+//!   The field name must be an alphabet followed by zero or more alphanumeric characters.
+//! * `.["field"]` if the field name contains special characters.
+//!   We recommend using a raw string literal for the query parameter (`#[query(r#"..."#)`).
 //! * `.[index]` for accessing an array element at position `index`.
 //!
 //! Note that mixing field access and index access at the same position of a document